requirements.txt
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
decorator==4.4.2
diffimg==0.3.0
idna==2.10
imutils==0.5.4
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
MouseInfo==0.1.3
mss==6.1.0
numpy==1.20.1
opencv-python==4.5.1.48
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.7
python-dateutil==2.8.1
PyTweening==1.0.3
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
xmltodict==0.12.0
